+++ cache/config.json
@@ -1,0 +1,6 @@
+{
+	"telemetry": {
+		"notifiedAt": "1755169292223",
+		"enabled": false
+	}
+}
