+++ tailwind.config.ts
@@ -9,24 +9,45 @@ export default {
     extend: {
       colors: {
         primary: {
-          DEFAULT: '#0f172a',
-          foreground: '#e2e8f0',
+          DEFAULT: 'hsl(var(--primary))',
+          foreground: 'hsl(var(--primary-foreground))',
         },
         accent: {
-          DEFAULT: '#00b894',
-          foreground: '#061a14',
+          DEFAULT: 'hsl(var(--accent))',
+          foreground: 'hsl(var(--accent-foreground))',
         },
         card: {
-          DEFAULT: '#0b1220',
-          foreground: '#cbd5e1',
+          DEFAULT: 'hsl(var(--card))',
+          foreground: 'hsl(var(--card-foreground))',
         },
         muted: {
-          DEFAULT: '#0a0f1a',
-          foreground: '#94a3b8',
+          DEFAULT: 'hsl(var(--muted))',
+          foreground: 'hsl(var(--muted-foreground))',
+        },
+        background: 'hsl(var(--background))',
+        foreground: 'hsl(var(--foreground))',
+        border: 'hsl(var(--border))',
+        input: 'hsl(var(--input))',
+        ring: 'hsl(var(--ring))',
+        
+        // Keep specific colors for gradients and special effects
+        emerald: {
+          400: '#34d399',
+        },
+        cyan: {
+          300: '#67e8f9',
+          400: '#22d3ee',
+        },
+        zinc: {
+          100: '#f4f4f5',
+          200: '#e4e4e7',
+          300: '#d4d4d8',
+          400: '#a1a1aa',
+          900: '#18181b',
         },
       },
       boxShadow: {
-        soft: '0 8px 24px rgba(0,0,0,0.25)'
+        soft: '0 8px 24px rgba(0,0,0,0.15)'
       }
     },
   },
